Split up fixture module into smaller files.
